chore: cleaning up code styling (#18223)
